rename Player/Screen to Player/PlayerScreen
