+++ app/player.tsx
@@ -1,4 +1,4 @@
-import { PlayerScreen } from '@/components/Player/Screen'
+import { PlayerScreen } from '@/components/Player/PlayerScreen'
 
 export default function Player () {
   return <PlayerScreen />
